requirements.txt
robotframework==7.0
robotframework-requests
robotframework-seleniumlibrary
selenium
webdriver-manager